Correct a 64-bit crash when calculating logarithmic colormaps of int32 data.
git-svn-id: https://svn.code.sf.net/p/pymca/code@721 a8702ac3-0063-49a2-b157-4a214673c9c1

+++ PyMca/sps/Src/sps_lut.c
@@ -291,10 +291,10 @@ void SPS_FindMinMax(void *data, int type, int cols, int rows,
      FINDMINMAX(float, FLT_MAX);
      break;
    case SPS_INT :
-     FINDMINMAX(int, LONG_MAX);
+     FINDMINMAX(int, INT_MAX);
      break;
    case SPS_UINT :
-     FINDMINMAX(unsigned int, ULONG_MAX);
+     FINDMINMAX(unsigned int, UINT_MAX);
      break;
    case SPS_SHORT :
      FINDMINMAX(short, SHRT_MAX);
